Refactor models Short -> Int, Biginteger -> Long Fix naming, adjust tests

+++ src/main/java/com/kambr/parser/onex/spo/dataClasses/FromAndPrice.kt
@@ -1,3 +1,3 @@
 package com.kambr.parser.onex.spo.dataClasses
 
-data class FromAndPrice(val from: Short, val price: Double)
+data class FromAndPrice(val from: Int, val price: Double)
