+++ src/main/java/com/kambr/parser/onex/sync/dataClasses/SpecialPriceOffer.kt
@@ -1,5 +1,3 @@
 package com.kambr.parser.onex.sync.dataClasses
 
-import java.math.BigDecimal
-
-data class SpecialPriceOffer(var fromSeats: Short, var price: BigDecimal)
+data class SpecialPriceOffer(var fromSeats: Int, var price: Double)
